fixed bugs in preprocessing
